move namespace and SequencerBlockData into their own files
