+++ crates/astria-sequencer-types/src/lib.rs
@@ -1,11 +1,13 @@
+pub(crate) mod namespace;
+pub(crate) mod sequencer_block_data;
 pub(crate) mod serde;
-pub mod types;
-pub mod utils;
+pub mod test_utils;
 
-pub use types::{
+pub use namespace::{
     Namespace,
-    SequencerBlockData,
+    DEFAULT_NAMESPACE,
 };
+pub use sequencer_block_data::SequencerBlockData;
 
 pub use crate::serde::{
     Base64Standard,
